requirements.txt
tkinter
numpy
Pillow